+++ pdf/lib/pdf.dart
@@ -26,6 +26,8 @@ import 'package:meta/meta.dart';
 import 'package:utf/utf.dart';
 import 'package:vector_math/vector_math_64.dart';
 
+import 'src/io.dart' if (dart.library.io) 'src/archive.dart';
+
 part 'src/annotation.dart';
 part 'src/array.dart';
 part 'src/ascii85.dart';