+++ pdf/lib/src/archive.dart
@@ -1,0 +1,5 @@
+import 'package:archive/archive.dart';
+
+import 'package:pdf/pdf.dart';
+
+DeflateCallback defaultDeflate =  ZLibEncoder().encode;