+++ pdf/lib/src/io.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+import 'package:pdf/pdf.dart';
+
+DeflateCallback defaultDeflate = zlib.encode;