+++ printing/lib/printing.dart
@@ -17,7 +17,6 @@
 library printing;
 
 import 'dart:async';
-import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui' as ui;